Updated some documentation.

+++ xsfftdemo.cpp
@@ -21,8 +21,8 @@ int main(int argc, char *argv[])
     xsComplex *dataIterator = data;
     for(long signalIndex = 0; signalIndex < dataLength; ++signalIndex, ++dataIterator) {
         //dataIterator->set((rand() % USHRT_MAX) - SHRT_MAX, 0.0);
-        //dataIterator->set(sin(2 * xsPI * signalIndex), 0.0);
-        dataIterator->set(signalIndex, 0.0);
+        //dataIterator->set(sin(2.0 * xsPI * (double)signalIndex), 0.0);
+        dataIterator->set((double)signalIndex, 0.0);
         cout << dataIterator->real() << "\n";
     }
